chore: remove unnecessary getters and dead code
