+++ contracts/src/c_pool/metadata.rs
@@ -122,31 +122,6 @@ pub fn write_total_weight(e: &Env, d: i128) {
     e.storage().instance().set(&key, &d)
 }
 
-//Read Token Share
-pub fn get_token_share(e: &Env) -> Address {
-    e.storage().persistent().bump(
-        &DataKey::TokenShare,
-        SHARED_LIFETIME_THRESHOLD,
-        SHARED_BUMP_AMOUNT,
-    );
-    e.storage()
-        .persistent()
-        .get::<DataKey, Address>(&DataKey::TokenShare)
-        .unwrap_optimized()
-}
-
-// Update Token Share
-pub fn put_token_share(e: &Env, contract_id: Address) {
-    e.storage()
-        .persistent()
-        .set(&DataKey::TokenShare, &contract_id);
-    e.storage().persistent().bump(
-        &DataKey::TokenShare,
-        SHARED_LIFETIME_THRESHOLD,
-        SHARED_BUMP_AMOUNT,
-    );
-}
-
 // Read Total Shares
 pub fn get_total_shares(e: &Env) -> i128 {
     e.storage().persistent().bump(
